Refactor course picker components and implement new planner filters
- Removed unused imports and components in CourseSearchContainer.
- Integrated SearchBox component for improved search functionality.
- Created PlannerFilters component to handle course filtering options.
- Added PlannerCourseListItem for displaying course details in the planner.
- Implemented PlannerSearchContainer to manage course search and filtering.
- Updated styling and structure for better responsiveness and usability.
- Ensured language settings are respected across components.
